test: verify the order of context functions
Add test cases to verify the order of context entries when applying a context function.

Since there is no direct way to test the order of the context entries, we use the `get entries()` function to return the entries in the same order as the context.

(cherry picked from commit 47f50656386cc15b0494d99a39eeb33ddbcb13ae)