fix: preserve the order inside a context
Fix an issue that changed the order in a given context. An expression should return the entries of a context in the same order as they are defined.

Refactor the evaluation context to avoid deep nesting of variable and function providers. The nesting is not necessary and hard to debug.

Change some function names of the context to make the code more explicit.

(cherry picked from commit f662f5d62975105d501a0e2ac4992ab5024c3266)