fix: Store context entries in defined order
The order of the context entries should be the same as defined in the context literal.

Ensure the order by using a SeqMap for the context entries. The usage of a SortedMap is not correct because this map sorts the entries alphabetic.

Add new test cases with non-alphabetic order to ensure that the context entries are not sorted.

(cherry picked from commit efcaee50d09f7aad8f3e13318020ba2121aab445)