feat: Create a new API with new result types
Currently, the engine API returns an Either as the result of an evaluation. The Either contains the result value if the valuation was successful. Otherwise, it contains the failure.

To expose the suppressed failures as part of the result, we need a new API. We can't change the existing API because of the backward-compatibility.

Instead of duplicating the existing API in the engine class, the new API moved into a separate class and package. This should keep the API and the engine more clean.

To expose and promote the new API, we need a new builder that returns the new API. The existing builder returns the old API.