fix: Parse unary-test with input comparison
Fix the parsing of a positive-unary-test expression. Before all expressions were parsed as UnaryTestExpression. As a result, the interpreter evaluated the unary-test (`< 5`) and if it returned false, it compared the input value with the result of the unary-test (i.e. `< 5` => `input < 5 or input = (input < 5)` => `null < 5 or null = (null < 5)` => `null or null = null` => `true`).

For positive-unary-test comparison with the input value (e.g. `< 5`), it makes no sense to compare the result of the comparison with the input value again.

Fix the behavior in the parser by handing these case separately and give them precedence over the general unary-tests behavior.