test: ignore unsupported null filter comparisons
Since we don't yet support list expressions with missing variables as
`null` values in that list, there is no reason to work on the filter
expressions for those list expressions.

Example, [1,2,x,4] should eventually evaluate to [1,2,null,4] but at
this time we evaluate it to an error stating that the variable is
missing. Testing that [1,2,x,4][item > 3] works correctly is therefore
not yet possible.

Co-authored-by: berkaycanbc <berkay.can@camunda.com>
Co-authored-by: Nicola Puppa <nicola.puppa@camunda.com>
Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>