test: no longer fail on incorrect filters
When the filter doesn't (always) return a boolean or a number (to filter
by index), then we no longer want to fail. Instead, we want to provide a
list of all the elements that did match the filter expression.

So, when the filter does not even evaluate to a boolean value, then we
want to product an empty list.

And, if the filter sometimes evaluates to a boolean value and sometimes
doesn't, then we want a list of those values for which the filter did
evaluate to the boolean `true`.

As both these cases are very similar, it makes sense to combine them
into a single test.

Co-authored-by: berkaycanbc <berkay.can@camunda.com>
Co-authored-by: Nicola Puppa <nicola.puppa@camunda.com>
Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>