test: Verify reported failures
If a function invocation returns null, we want to verify that a failure is reported.

A user of the API can access the evaluation failures to understand why an evaluation returned null. So, we should verify the reported failures.

To verify the failures, we need a new base test class that returns the whole evaluation result, instead of only the return value. Create a new test matcher to increase the readability of the tests and better failure messages.