build: Fix the licence Maven plugin
Move the configuration of the license plugin into the plugin management section to override the parent POM.

Create two profiles: one profile for adding license headers and apply formatting, and another profile for only checking if license headers are present.

The second profile is used in the CI. The first profile is used by default for local development.