fix: Return false if the value is null
Change the behavior of the function `is defined()` to return `true` if the value is not `null`.

Previously, the function returned `true` if the value was `null`. It returned `false` if the value was a non-existing variable or context entry (i.e. a ValError).

The changes in the null-handling of the FEEL engine to replace a non-exising variable or context entry with `null` made this function useless. It returned always `true`.

A complete fix of the function is not possible because of the new null-handling. But this change in the semantics of the function reduces the impact for the users (i.e. the regression).

The function still works as before for non-existing variables and context entries. But it may break for variables and context entries with a `null` value.