fix: Invoke context() with custom context
The context function expects a list argument with context entries that contain a key and a value. The function checks that the key value is a string. To support a custom context value, the value must be resolved via value-mapper first before checking the type. Otherwise, the value may be not a FEEL Val type but something else.

(cherry picked from commit 2416531741a7ed592194b7b533fc05e49f155768)