fix: Return fatal error in operations
Adjust all operations to return a fatal error instead of returning null. This is required to limit the access to the special input symbol `?` outside an unary-tests.

Restructure some helper methods to restrict the handling of fatal errors to a few places. This should help to reduce possible problems with fatal errors.

Move some methods to group different aspects together.