fix: Ignore whitespace characters
Adjust the whitespace handling of the parser to ignore FEEL-specific whitespaces as defined in the DMN specification.

The JavaWhitespace class worked very well because it supported also Java-like comments, as required by the DMN specification. Instead of adding additional whitespace rules for the individual parsers, extend the JavaWhitespace class by making a copy of it. This is less error-prone, easier to maintain, and more efficient.