perf: fix O(N^2) performance of appending suppressedFailures
Appending a failure to a List is a O(N) operation: doing it N times
becomes O(N^2).
ListBuffer is a mutable builder that appends elements in O(N) and
creates a list in O(1).

Note that this class was already mutable (because of the var), but now
uses an mutable data structure.

(cherry picked from commit 3bd2b95cc9c12090527179ce94e16ca309e83c16)